Fix test failures
