+++ components/org.wso2.carbon.si.metrics.core/src/test/java/org/wso2/carbon/si/metrics/core/StatisticsTestCase.java
@@ -118,8 +118,10 @@ public class StatisticsTestCase {
         AssertJUnit.assertEquals("test.size", bufferedEventsTracker.getName(eventBufferHolder));
     }
     
-    @Test
+    @Test (enabled = false)
     public void statisticsTest1() throws InterruptedException {
+        // This test case has been disabled since it has a issue on enabling/disabling metrics at the Siddhi side.
+        // Issue: https://github.com/wso2/carbon-analytics/issues/1709
         log.info("statistics test 1");
         SiddhiManager siddhiManager = new SiddhiManager();
         StatisticsConfiguration statisticsConfiguration = new StatisticsConfiguration(new SPMetricsFactory());
