+++ components/org.wso2.carbon.streaming.integrator.core/src/test/java/org/wso2/carbon/streaming/integrator/core/HACoordinationRecordTableHandlerTest.java
