+++ components/org.wso2.carbon.streaming.integrator.core/src/test/resources/testng.xml
@@ -22,11 +22,11 @@ under the License.
 <suite name="wso2sp-test-suite">
     <test name="stream-processor-core" parallel="false">
         <classes>
-            <!--<class name="org.wso2.carbon.streaming.integrator.core.HACoordinationSinkHandlerTest"/>-->
+            <!--<class name="HACoordinationSinkHandlerTest"/>-->
             <class name="org.wso2.carbon.streaming.integrator.core.HACoordinationSourceHandlerTest"/>
             <!--<class name="org.wso2.carbon.streaming.integrator.core.HACoordinationRecordTableHandlerTest"/>-->
             <class name="org.wso2.carbon.streaming.integrator.core.DynamicHtmlGenTest"/>
-            <!--<class name="org.wso2.carbon.streaming.integrator.core.BeanTest"/>-->
+            <!--<class name="BeanTest"/>-->
             <class name="org.wso2.carbon.streaming.integrator.core.DBPersistenceStoreTest"/>
         </classes>
     </test>
