+++ pom.xml
@@ -1549,8 +1549,8 @@
 
     <properties>
         <carbon.analytics.version>3.0.0-SNAPSHOT</carbon.analytics.version>
-        <siddhi.version>5.0.3-SNAPSHOT</siddhi.version>
-        <siddhi.bundle.version>5.0.3-SNAPSHOT</siddhi.bundle.version>
+        <siddhi.version>5.1.1-SNAPSHOT</siddhi.version>
+        <siddhi.bundle.version>5.1.1-SNAPSHOT</siddhi.bundle.version>
         <siddhi.version.range>[5.0.0, 6.0.0)</siddhi.version.range>
         <carbon.analytics-common.version>6.1.26</carbon.analytics-common.version>
         <carbon.analytics-common.version.range>[6.0.66, 6.2.0)</carbon.analytics-common.version.range>
